# PERF: DataFrame.loc and Series.loc when key is a MultiIndex