- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.2.0.rst` file if fixing a bug or adding a new feature.

Avoid hitting `MultiIndex.values` in `is_bool_indexer`. 

```
from asv_bench.benchmarks.indexing import MultiIndexing
b = MultiIndexing()
b.setup(True)
%timeit b.time_loc_multiindex(True)

# 53.3 ms ± 2.18 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)   # <- main
# 28.6 ms ± 821 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- PR
```